pyproject.toml
[tool.poetry]
name = "deal-droid"
version = "0.1.0"
description = "Sales chatbot"
authors = ["Pee Tankulrat <8246751+PiiTan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.33.0"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.4"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.3.1"
langchain = "^0.2.6"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
